breaking out query code
